+++ Sources/APIServer/Utilities/DatabaseConfiguration.swift
@@ -204,4 +204,5 @@ func registerMigrations(on app: Application) {
     app.migrations.add(AddTestSetupLastRetestedManifestHash())
     app.migrations.add(CreatePreEnrollments())
     app.migrations.add(SessionRecord.migration)
+    app.migrations.add(AddUserLastSeenAt())
 }